s-nerfpp/requirements.txt
torch==2.0.1
absl_py==1.4.0
accelerate==0.22.0
albumentations==0.5.2
braceexpand==0.1.7
dearpygui==1.11.1
easydict==1.13
gin_config==0.5.0
glfw==2.7.0
imageio==2.31.2
imgaug==0.4.0
joblib==1.3.2
kornia==0.6.10
matplotlib==3.5.3
mediapy==1.2.0
memory_profiler==0.61.0
mmcv==2.1.0
numpy==1.25.2
nuscenes_devkit==1.1.11
omegaconf==2.3.0
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
opencv_python_headless==4.9.0.80
pandas==2.2.1
Pillow==10.0.0
plyfile==1.0.3
PyOpenGL==3.1.7
pyquaternion==0.9.9
pytorch_lightning==2.2.1
PyYAML==6.0.1
PyYAML==6.0.1
rawpy==0.18.1
scikit_learn==1.4.1.post1
scipy==1.13.0
setuptools==65.5.1
six==1.16.0
scikit-image
sync_batchnorm==0.0.1
tensorboardX==2.6.2.2
torch_efficient_distloss==0.1.3
torch_scatter==2.1.1
torchvision==0.15.2
tqdm==4.66.1
trimesh==3.23.5
wandb==0.16.5
webdataset==0.2.86
rtree
